Code cleanup: - Removed redundant String.substring() call in DebugToStringStyle - Removed redundant parameter from String.substring() call in Player - Removed redundant cast to Quaffable in ActionQuaffEntity - Removed redundant cast to int in PotionType - Removed redundant return in StateApproachComestible - Removed redundant null check in EntityArrow - Removed redundant null check in TreeNode - Removed redundant mouseMoved = false in GameInputProcessor - Removed redundant field initialisation in AIState.duration - Removed redundant field initialisation in NutritionState.importance - Removed redundant field initialisation in TileType.lightIntensity - Removed redundant field initialisation in TreeNode.identityHashCode - Removed two redundant contains checks from AStarPathfinder - Added a TODO in GameWidget.GameWidgetClickListener - Replaced Optional.isPresent() with Optional.orElse(null) in Item.copy() - Replaced Optional.isPresent() with Optional.orElse(null) in PlayerItemVisitor - Used StandardCharsets.UTF_8 in Dungeon - Removed duplicate deleteSave() code in Dungeon - Replaced switch statement with if statement for the sake of brevity in PlayerDefaultVisitors - Suppressed MethodDoesntCallSuperMethod in Word.clone() - Replaced Map.entrySet().forEach() with Map.forEach() in Container - Used Map.getOrDefault() in Attributes - Used expression lambda in EntityLiving - Used expression lambdas in Wishes - Added @param description to level in AStarPathfinder.getHeuristicCost javadoc - Fixed typo in Dungeon.prompt javadoc - Referred to Entity#isBeingRemoved() instead of Entity#beingRemoved in EntityStore.removeEntity javadoc - Replaced Stream().filter().count() > 0 with Stream.anyMatch() in Path - Separated string concatenations within StringBuilder.append() in DebugToStringStyle - Replaced string concatenations with StringBuilder in Colour - Replaced string concatenations with StringBuilder in Prompt - Fixed key/values in ItemSpellbook.spellLevelMap being inverted. TODO: verify this was intentional

+++ src/main/java/jr/dungeon/items/quaffable/potions/PotionType.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public enum PotionType {
-	POTION_HEALTH((ent, potency) -> ent.heal((int) Math.round(potency)));
+	POTION_HEALTH((ent, potency) -> ent.heal(Math.round(potency)));
 	
 	private final PotionEffect effect;
 }
